Crate stable_mir

The WIP stable interface to rustc internals.

For more information see https://github.com/rust-lang/project-stable-mir

Note

This API is still completely unstable and subject to change.

This crate shall contain all type definitions and APIs that we expect third-party tools to invoke to interact with the compiler.

The goal is to eventually be published on crates.io.

Modules

fold

mir

ty

visitor

Structs

AllocId

A unique identification number for each provenance

Crate

Holds information about a crate.

CrateItem

Holds information about an item in the crate. For now, it only stores the item DefId. Use functions inside rustc_internal module to use this item.

DefId

A unique identification number for each item accessible for the current compilation unit.

Opaque

A type that provides internal information but that can still be used for debug purpose.

Enums

CompilerError

An error type used to represent an error that has already been reported by the compiler.

Statics

TLV 🔒

Traits

Context

Functions

all_local_items

Retrieve all items in the local crate that have a MIR associated with them.

all_trait_decls

all_trait_impls

entry_fn

Return the function where execution starts if the current crate defines that. This is usually main, but could be start if the crate is a no-std crate.

external_crates

Try to find a crate with the given name.

find_crate

Try to find a crate with the given name.

local_crate

Access to the local crate.

opaque

run

trait_decl

trait_impl

with

Loads the current context and calls a function with it. Do not nest these, as that will ICE.

Type Aliases

CrateItems

A list of crate items.

CrateNum

The number that identifies a crate.

DefKind

ImplTraitDecls

A list of impl trait decls.

Symbol

Use String for now but we should replace it.

TraitDecls

A list of trait decls.