stable_mir
Module abi

Structs

AddressSpace

An identifier that specifies the address space that some operation should operate on. Special address spaces have an effect on code generation, depending on the target and the address spaces it implements.

ArgAbi

Information about the ABI of a function’s argument, or return value.

FnAbi

A function ABI definition.

Layout

LayoutShape

The layout of a type in memory.

TyAndLayout

The layout of a type, alongside the type itself.

WrappingRange

Inclusive wrap-around range of valid values (bitwise representation), that is, if start > end, it represents start..=MAX, followed by 0..=end.

Enums

CallConvention

General language calling conventions.

FieldsShape

Describes how the fields of a type are shaped in memory.

FloatLength

Enum representing the existing float lengths.

IntegerLength

Enum representing the existing integer lengths.

PassMode

How a function argument should be passed in to the target function.

Primitive

Fundamental unit of memory access and layout.

Scalar

Information about one scalar component of a Rust type.

TagEncoding

ValueAbi

Describes how values of the type are passed by target ABIs, in terms of categories of C types there are ABI rules for.

VariantsShape