stable_mir::abi
Enum FieldsShape

pub enum FieldsShape {
    Primitive,
    Union(NonZero<usize>),
    Array {
        stride: MachineSize,
        count: u64,
    },
    Arbitrary {
        offsets: Vec<MachineSize>,
    },
}

Describes how the fields of a type are shaped in memory.

Variants

Primitive

Scalar primitives and !, which never have fields.

Union(NonZero<usize>)

All fields start at no offset. The usize is the field count.

Array

Array/vector-like placement, with all fields of identical types.

Fields

stride: MachineSize

count: u64

Arbitrary

Struct-like placement, with precomputed offsets.

Fields are guaranteed to not overlap, but note that gaps before, between and after all the fields are NOT always padding, and as such their contents may not be discarded. For example, enum variants leave a gap at the start, where the discriminant field in the enum layout goes.

Fields

offsets: Vec<MachineSize>

Offsets for the first byte of each field, ordered to match the source definition order. I.e.: It follows the same order as crate::ty::VariantDef::fields(). This vector does not go in increasing order.

Implementations

impl FieldsShape

pub fn fields_by_offset_order(&self) -> Vec<FieldIdx>

pub fn count(&self) -> usize