Module stable_mir::compiler_interface

Define the interface with the Rust compiler.

StableMIR users should not use any of the items in this module directly. These APIs have no stability guarantee.

Statics

TLV 🔒

Traits

Context

This trait defines the interface between stable_mir and the Rust compiler. Do not use this directly.

Functions

run

with 🔒

Execute the given function with access the compiler Context.