stable_mir::compiler_interface
Function with

pub(crate) fn with<R>(f: impl FnOnce(&dyn Context) -> R) -> R

Execute the given function with access the compiler Context.

I.e., This function will load the current context and calls a function with it. Do not nest these, as that will ICE.