stable_mir
Module error

When things go wrong, we need some error handling. There are a few different types of errors in StableMIR:

CompilerError: This represents errors that can be raised when invoking the compiler.

Error: Generic error that represents the reason why a request that could not be fulfilled.

Macros

error 🔒

Structs

Error

A generic error to represent an API request that cannot be fulfilled.

Enums

CompilerError

An error type used to represent an error that has already been reported by the compiler.