stable_mir
Module mir

Re-exports

pub use visit::MirVisitor;

Modules

alloc

This module provides methods to retrieve allocation information, such as static variables.

body ๐

mono

pretty

visit

The Stable MIR Visitor

Structs

BasicBlock

Body

The SMIR representation of a single function.

ConstOperand

CopyNonOverlapping

InlineAsmOperand

LocalDecl

Place

SourceInfo

Statement

SwitchTargets

The possible branch sites of a TerminatorKind::SwitchInt.

Terminator

UserTypeProjection

VarDebugInfo

Debug information pertaining to a user variable.

VarDebugInfoFragment

Enums

AggregateKind

AssertMessage

BinOp

BorrowKind

CastKind

CoroutineDesugaring

CoroutineKind

CoroutineSource

FakeBorrowKind

FakeReadCause

The FakeReadCause describes the type of pattern why a FakeRead statement exists.

MutBorrowKind

Mutability

NonDivergingIntrinsic

NullOp

Operand

PointerCoercion

ProjectionElem

RetagKind

Describes what kind of retag is to be performed

Rvalue

Safety

StatementKind

TerminatorKind

UnOp

UnwindAction

VarDebugInfoContents

Variance

Constants

RETURN_LOCAL

Type Aliases

BasicBlockIdx

FieldIdx

The source-order index of a field in a variant.

Local

SourceScope

Successors