stable_mir::mir
Module body

Structs

BasicBlock

Body

The SMIR representation of a single function.

ConstOperand

CopyNonOverlapping

InlineAsmOperand

LocalDecl

Place

SourceInfo

Statement

SwitchTargets

The possible branch sites of a TerminatorKind::SwitchInt.

Terminator

UserTypeProjection

VarDebugInfo

Debug information pertaining to a user variable.

VarDebugInfoFragment

Enums

AggregateKind

AssertMessage

BinOp

BorrowKind

CastKind

CoroutineDesugaring

CoroutineKind

CoroutineSource

FakeBorrowKind

FakeReadCause

The FakeReadCause describes the type of pattern why a FakeRead statement exists.

MutBorrowKind

Mutability

NonDivergingIntrinsic

NullOp

Operand

PointerCoercion

ProjectionElem

RetagKind

Describes what kind of retag is to be performed

Rvalue

Safety

StatementKind

TerminatorKind

UnOp

UnwindAction

VarDebugInfoContents

Variance

Constants

RETURN_LOCAL

Type Aliases

BasicBlockIdx

Coverage 🔒

The rustc coverage data structures are heavily tied to internal details of the coverage implementation that are likely to change, and are unlikely to be useful to third-party tools for the foreseeable future.

FieldIdx

The source-order index of a field in a variant.

Local

LocalDecls 🔒

LocalDefId 🔒

SourceScope

Successors

UserTypeAnnotationIndex 🔒