stable_mir::mir::mono
Struct Instance

pub struct Instance {
    pub kind: InstanceKind,
    pub def: InstanceDef,
}

Fields
kind: InstanceKind

The type of instance.

def: InstanceDef

An ID used to get the instance definition from the compiler. Do not use this field directly.

Implementations

impl Instance

pub fn args(&self) -> GenericArgs

Get the arguments this instance was instantiated with.

pub fn body(&self) -> Option<Body>

Get the body of an Instance.

The body will be eagerly monomorphized and all constants will already be evaluated.

This method will return the intrinsic fallback body if one was defined.

pub fn has_body(&self) -> bool

Check whether this instance has a body available.

For intrinsics with fallback body, this will return true. It is up to the user to decide whether to specialize the intrinsic or to use its fallback body.

For more information on fallback body, see https://github.com/rust-lang/rust/issues/93145.

This call is much cheaper than instance.body().is_some(), since it doesn’t try to build the StableMIR body.

pub fn is_foreign_item(&self) -> bool

pub fn ty(&self) -> Ty

Get the instance type with generic instantiations applied and lifetimes erased.

pub fn fn_abi(&self) -> Result<FnAbi, Error>

Retrieve information about this instance binary interface.

pub fn mangled_name(&self) -> Symbol

Retrieve the instance’s mangled name used for calling the given instance.

This will also look up the correct name of instances from upstream crates.

pub fn name(&self) -> Symbol

Retrieve the instance name for diagnostic messages.

This will return the specialized name, e.g., std::vec::Vec<u8>::new.

pub fn trimmed_name(&self) -> Symbol

Return a trimmed name of the given instance including its args.

If a symbol name can only be imported from one place for a type, and as long as it was not glob-imported anywhere in the current crate, we trim its path and print only the name.

pub fn intrinsic_name(&self) -> Option<Symbol>

Retrieve the plain intrinsic name of an instance if it’s an intrinsic.

The plain name does not include type arguments (as trimmed_name does), which is more convenient to match with intrinsic symbols.

pub fn resolve(def: FnDef, args: &GenericArgs) -> Result<Instance, Error>

Resolve an instance starting from a function definition and generic arguments.

pub fn resolve_drop_in_place(ty: Ty) -> Instance

Resolve the drop in place for a given type.

pub fn resolve_for_fn_ptr( def: FnDef, args: &GenericArgs, ) -> Result<Instance, Error>

Resolve an instance for a given function pointer.

pub fn resolve_closure( def: ClosureDef, args: &GenericArgs, kind: ClosureKind, ) -> Result<Instance, Error>

Resolve a closure with the expected kind.

pub fn is_empty_shim(&self) -> bool

Check whether this instance is an empty shim.

Allow users to check if this shim can be ignored when called directly.

We have decided not to export different types of Shims to StableMIR users, however, this is a query that can be very helpful for users when processing DropGlue.

When generating code for a Drop terminator, users can ignore an empty drop glue. These shims are only needed to generate a valid Drop call done via VTable.

pub fn try_const_eval(&self, const_ty: Ty) -> Result<Allocation, Error>

Try to constant evaluate the instance into a constant with the given type.

This can be used to retrieve a constant that represents an intrinsic return such as type_id.

pub fn emit_mir<W: Write>(&self, w: &mut W) -> Result<()>

Emit the body of this instance if it has one.