stable_mir
Module ty

Macros

index_impl 🔒

Structs

AdtDef

AliasDef

AliasTerm

AliasTy

Allocation

Binder

A binder represents a possibly generic type and its bound vars.

BoundRegion

BoundTy

BrNamedDef

ClosureDef

CoercePredicate

ConstDef

CoroutineDef

CoroutineWitnessDef

EarlyBinder

EarlyParamRegion

ExistentialProjection

ExistentialTraitRef

An existential reference to a trait where Self is not included.

FieldDef

FnDef

Hold information about a function definition in a crate.

FnSig

ForeignDef

Hold information about a ForeignItem in a crate.

ForeignModule

ForeignModuleDef

GenericArgs

A list of generic arguments.

GenericDef

GenericParamDef

GenericPredicates

Generics

ImplDef

A trait impl definition.

IntrinsicDef

LineInfo

Information you get from Span in a struct form. Line and col start from 1.

MirConst

Represents a constant in MIR

MirConstId

OutlivesPredicate

ParamConst

ParamDef

ParamTy

Placeholder

ProjectionPredicate

Prov

ProvenanceMap

Stores the provenance information of pointers stored in memory.

Region

RegionDef

Span

SubtypePredicate

TraitDecl

TraitDef

A trait’s definition.

TraitPredicate

TraitRef

A complete reference to a trait, i.e., one where Self is known.

Ty

TyConst

Represents a constant in the type system

TyConstId

TypeAndMut

UnevaluatedConst

VariantDef

Definition of a variant, which can be either a struct / union field or an enum variant.

VariantIdx

The source-order index of a variant in a type.

Enums

Abi

AdtKind

AliasKind

AliasRelationDirection

BoundRegionKind

BoundTyKind

BoundVariableKind

ClauseKind

ClosureKind

ConstantKind

DynKind

ExistentialPredicate

FloatTy

ForeignItemKind

GenericArgKind

GenericParamDefKind

ImplPolarity

IntTy

Movability

Pattern

Represents a pattern in the type system

PredicateKind

PredicatePolarity

RegionKind

RigidTy

TermKind

TraitSpecializationKind

TyConstKind

TyKind

UintTy

Traits

IndexedVal

Type Aliases

Align

BoundVar 🔒

Bytes

DebruijnIndex 🔒

Ident 🔒

ImplTrait

InitMaskMaterialized

PolyFnSig

Promoted

RegionOutlivesPredicate

Size

Size in bytes.

TypeOutlivesPredicate

UniverseIndex 🔒