Module stable_mir::mir

Re-exports

pub use visit::MirVisitor;

Modules

alloc

This module provides methods to retrieve allocation information, such as static variables.

body 🔒

mono

pretty

visit

The Stable MIR Visitor

Structs

BasicBlock

Body

The SMIR representation of a single function.

ConstOperand

Constant

CopyNonOverlapping

InlineAsmOperand

LocalDecl

Place

SourceInfo

Statement

SwitchTargets

The possible branch sites of a TerminatorKind::SwitchInt.

Terminator

UserTypeProjection

VarDebugInfo

Debug information pertaining to a user variable.

VarDebugInfoFragment

Enums

AggregateKind

AssertMessage

BinOp

BorrowKind

CastKind

CoroutineDesugaring

CoroutineKind

CoroutineSource

FakeReadCause

The FakeReadCause describes the type of pattern why a FakeRead statement exists.

MutBorrowKind

Mutability

NonDivergingIntrinsic

NullOp

Operand

PointerCoercion

ProjectionElem

RetagKind

Describes what kind of retag is to be performed

Rvalue

Safety

StatementKind

TerminatorKind

UbKind

UnOp

UnwindAction

VarDebugInfoContents

Variance

Constants

RETURN_LOCAL

Type Aliases

BasicBlockIdx

FieldIdx

The source-order index of a field in a variant.

Local

SourceScope

Successors