stable_mir::ty
Enum TyKind

pub enum TyKind {
    RigidTy(RigidTy),
    Alias(AliasKind, AliasTy),
    Param(ParamTy),
    Bound(usize, BoundTy),
}

Variants

RigidTy(RigidTy)

Alias(AliasKind, AliasTy)

Param(ParamTy)

Bound(usize, BoundTy)

Implementations

impl TyKind

pub fn rigid(&self) -> Option<&RigidTy>

pub fn is_unit(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_char(&self) -> bool

pub fn is_trait(&self) -> bool

pub fn is_enum(&self) -> bool

pub fn is_struct(&self) -> bool

pub fn is_union(&self) -> bool

pub fn is_adt(&self) -> bool

pub fn is_ref(&self) -> bool

pub fn is_fn(&self) -> bool

pub fn is_fn_ptr(&self) -> bool

pub fn is_primitive(&self) -> bool

pub fn is_float(&self) -> bool

pub fn is_integral(&self) -> bool

pub fn is_numeric(&self) -> bool

pub fn is_signed(&self) -> bool

pub fn is_str(&self) -> bool

pub fn is_cstr(&self) -> bool

pub fn is_slice(&self) -> bool

pub fn is_array(&self) -> bool

pub fn is_mutable_ptr(&self) -> bool

pub fn is_raw_ptr(&self) -> bool

pub fn is_any_ptr(&self) -> bool

Tests if this is any kind of primitive pointer type (reference, raw pointer, fn pointer).

pub fn is_coroutine(&self) -> bool

pub fn is_closure(&self) -> bool

pub fn is_box(&self) -> bool

pub fn is_simd(&self) -> bool

pub fn trait_principal(&self) -> Option<Binder<ExistentialTraitRef>>

pub fn builtin_index(&self) -> Option<Ty>

Returns the type of ty[i] for builtin types.

pub fn builtin_deref(&self, explicit: bool) -> Option<TypeAndMut>

Returns the type and mutability of *ty for builtin types.

The parameter explicit indicates if this is an explicit dereference. Some types โ notably unsafe ptrs โ can only be dereferenced explicitly.

pub fn fn_sig(&self) -> Option<PolyFnSig>

Get the function signature for function like types (Fn, FnPtr, and Closure)

pub fn discriminant_ty(&self) -> Option<Ty>

Get the discriminant type for this type.

pub fn fn_def(&self) -> Option<(FnDef, &GenericArgs)>

Deconstruct a function type if this is one.